geolocation/position-coords-002.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: the attribute coords of interface Position is readonly</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      var t = async_test("Test checks that the attribute coords of interface Position is readonly");
      setup({timeout: 20000});
      try {
        navigator.geolocation.getCurrentPosition(successCallback, errorCallback);
      } catch (err) {
        t.step(function () {
          assert_unreached("Error message: " + err.message);
        });
        t.done();
      }
      function successCallback (position) {
        t.step(function () {
          position.coords = "tests";
          assert_false(position.coords == "tests", "the attribute coords of interface Position is readonly");
        });
        t.done();
      }
      function errorCallback (error) {
        t.step(function () {
          assert_unreached("Error message: " + error.message);
        });
        t.done();
      }
    </script>
    <div id="log"></div>
  </body>
</html>
